# Conflicts:
#   src/main/java/ru/itis/studentsgiftery/controllers/BrandsController.java
#   src/main/java/ru/itis/studentsgiftery/controllers/OrganizationsController.java

+++ src/main/java/ru/itis/studentsgiftery/services/BalanceService.java
@@ -1,11 +1,10 @@
 package ru.itis.studentsgiftery.services;
 
 import ru.itis.studentsgiftery.dto.AccountDto;
-import ru.itis.studentsgiftery.exceptions.LowBalanceException;
 import ru.itis.studentsgiftery.models.Account;
 
 public interface BalanceService {
-    Integer getAccountBalance(Long accountId);
+    Long getAccountBalance(Long accountId);
 
     AccountDto addBalance(Long accountId, Integer amount);
